requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
autopep8==2.0.2
backcall==0.2.0
backports.functools-lru-cache==1.6.5
boto3==1.28.57
botocore==1.31.57
certifi==2023.7.22
charset-normalizer==3.3.0
cmake==3.27.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
filelock==3.12.4
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.31
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
ipykernel==6.25.2
ipython==8.12.0
ipywidgets==8.1.1
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jupyter-client==7.3.4
jupyter_core==5.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lit==17.0.1
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.0
requests==2.31.0
s3transfer==0.7.0
seaborn==0.13.0
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.1
tqdm==4.66.1
traitlets==5.10.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.13.11
wcwidth==0.2.7
wheel==0.41.2
widgetsnbextension==4.0.9
zipp==3.15.0
timm==0.9.7
scikit-learn==1.3.1
